Fix to allow `*` attributes too if specific also specified
The attributes schema allows both specific tag names (say, `a`) and a
generic catch all (`*`).
Previously, if a strict specific schema was allowed (such as that only
`/^language-./` classes are allowed on `code`), and someone allowed all
classes on the `*` wildcard, then this latter intent was not honored.
Now, it works, either works.

Closes GH-27.
Closes GH-28.